apps/backend/pyproject.toml
[project]
name = "val-cat-conjugation-backend"
version = "0.0.1"
description = "A backend for valencian/catalan verb conjugation."
authors = [
    {name = "pavelbeard", email = "pavelborodin1995@gmail.com"},
]
dependencies = ["fastapi[standard]>=0.115.14", "fastapi>=0.115.14", "openai>=1.93.0", "pydantic-settings>=2.10.1", "pymongo[srv]>=4.13.2", "beautifulsoup4>=4.13.4", "lxml>=6.0.0", "pytest>=8.4.1", "pytest-asyncio>=1.0.0", "ollama>=0.5.1", "pydantic-ai-slim[openai]>=0.4.2", "pytest-mock>=3.14.1", "flake8>=7.3.0"]
requires-python = ">=3.13"
readme = "README.md"
license = {text = "Free to use"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[pytest]
mock_use_standalone_module = true
